+++ integrations/chroma/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [integrations/chroma-v4.0.0] - 2026-01-12
+
+### 🚀 Features
+
+- Add `client_settings` to `ChromaDocumentStore` (#2651)
+
+### 📚 Documentation
+
+- Fix: doc links (#2662)
+
+### 🧹 Chores
+
+- Make fmt command more forgiving (#2671)
+- [**breaking**] Chroma - drop Python 3.9 and use X|Y typing (#2701)
+
+### 🌀 Miscellaneous
+
+- Fix: Fix doc links (#2661)
+
 ## [integrations/chroma-v3.5.0] - 2026-01-07
 
 ### 🚀 Features